+++ src/lib/utils/thorchain/index.ts
@@ -1,23 +1,28 @@
 import type { AccountId } from '@shapeshiftoss/caip'
+import { type AssetId, bchChainId, fromAccountId, fromAssetId } from '@shapeshiftoss/caip'
+import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
 import type { QueryClient } from '@tanstack/react-query'
+import axios from 'axios'
+import { getConfig } from 'config'
+import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import { queryClient } from 'context/QueryClientProvider/queryClient'
 import type { Asset } from 'lib/asset-service'
-import { bnOrZero } from 'lib/bignumber/bignumber'
+import type { BigNumber, BN } from 'lib/bignumber/bignumber'
+import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit, toBaseUnit } from 'lib/math'
 import { poll } from 'lib/poll/poll'
-import {
-  type EstimatedFeesQueryKey,
-  queryFn as getEstimatedFeesQueryFn,
-} from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
+import type { ThornodeStatusResponse } from 'lib/swapper/swappers/ThorchainSwapper/types'
+import type { EstimatedFeesQueryKey } from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
+import { queryFn as getEstimatedFeesQueryFn } from 'pages/Lending/hooks/useGetEstimatedFeesQuery'
 import type { IsSweepNeededQueryKey } from 'pages/Lending/hooks/useIsSweepNeededQuery'
 import { queryFn as isSweepNeededQueryFn } from 'pages/Lending/hooks/useIsSweepNeededQuery'
 import { selectPortfolioCryptoBalanceBaseUnitByFilter } from 'state/slices/common-selectors'
+import type { getThorchainLendingPosition } from 'state/slices/opportunitiesSlice/resolvers/thorchainLending/utils'
 import type { ThorchainSaversWithdrawQuoteResponseSuccess } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/types'
-import {
-  fromThorBaseUnit,
-  getThorchainTransactionStatus,
-} from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
+import type { getThorchainSaversPosition } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
+import { THOR_PRECISION } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
+import type { AccountMetadata } from 'state/slices/portfolioSlice/portfolioSliceCommon'
 import { isUtxoChainId } from 'state/slices/portfolioSlice/utils'
 import { selectMarketDataById } from 'state/slices/selectors'
 import { store } from 'state/store'
@@ -28,6 +33,29 @@ import {
   type GetThorchainSaversWithdrawQuoteQueryKey,
   queryFn as getThorchainSaversWithdrawQuoteQueryFn,
 } from './hooks/useGetThorchainSaversWithdrawQuoteQuery'
+
+const getThorchainTransactionStatus = async (txHash: string, skipOutbound?: boolean) => {
+  const thorTxHash = txHash.replace(/^0x/, '')
+  const { data: thorTxData, status } = await axios.get<ThornodeStatusResponse>(
+    `${getConfig().REACT_APP_THORCHAIN_NODE_URL}/lcd/thorchain/tx/status/${thorTxHash}`,
+    // We don't want to throw on 404s, we're parsing these ourselves
+    { validateStatus: () => true },
+  )
+
+  if ('error' in thorTxData || status === 404) return TxStatus.Unknown
+  // Tx has been observed, but swap/outbound Tx hasn't been completed yet
+  if (
+    // Despite the Tx being observed, things may be slow to be picked on the THOR node side of things i.e for swaps to/from BTC
+    thorTxData.stages.inbound_finalised?.completed === false ||
+    thorTxData.stages.swap_status?.pending === true ||
+    (!skipOutbound && thorTxData.stages.outbound_signed?.completed === false)
+  )
+    return TxStatus.Pending
+  if (thorTxData.stages.swap_status?.pending === false) return TxStatus.Confirmed
+
+  // We shouldn't end up here, but just in case
+  return TxStatus.Unknown
+}
 
 export const waitForThorchainUpdate = ({
   txHash,
@@ -253,4 +281,68 @@ export const fetchHasEnoughBalanceForTxPlusFeesPlusSweep = async ({
     bnOrZero(balanceCryptoBaseUnit).gt(0) && hasEnoughBalance
 
   return { hasEnoughBalance: hasEnoughBalanceForTxPlusFeesPlusSweep, missingFunds }
+}
+export const fromThorBaseUnit = (valueThorBaseUnit: BigNumber.Value | null | undefined): BN =>
+  bnOrZero(valueThorBaseUnit).div(bn(10).pow(THOR_PRECISION)) // to crypto precision from THOR 8 dp base unit
+
+export const toThorBaseUnit = ({
+  valueCryptoBaseUnit,
+  asset,
+}: {
+  valueCryptoBaseUnit: BigNumber.Value | null | undefined
+  asset: Asset
+}): BN => {
+  if (!asset?.precision) return bn(0)
+
+  return bnOrZero(valueCryptoBaseUnit)
+    .div(bn(10).pow(asset?.precision)) // to crypto precision from THOR 8 dp base unit
+    .times(bn(10).pow(THOR_PRECISION))
+    .decimalPlaces(0) // THORChain expects ints, not floats
+}
+
+export const getThorchainFromAddress = async ({
+  accountId,
+  assetId,
+  getPosition,
+  accountMetadata,
+  wallet,
+}: {
+  accountId: AccountId
+  assetId: AssetId
+  getPosition: typeof getThorchainLendingPosition | typeof getThorchainSaversPosition
+  accountMetadata: AccountMetadata
+  wallet: HDWallet
+}): Promise<string> => {
+  const { chainId } = fromAssetId(assetId)
+  if (!isUtxoChainId(chainId)) return Promise.resolve(fromAccountId(accountId).account)
+
+  try {
+    const position = await getPosition({
+      accountId,
+      assetId,
+    })
+    if (!position) throw new Error(`No position found for assetId: ${assetId}`)
+    const address: string = (() => {
+      // THORChain lending position
+      if ('owner' in position) return position.owner
+      // THORChain savers position
+      if ('asset_address' in position) return position.asset_address
+      // For type completeness - if we have a response, we *should* either have an `owner` or `asset_address` property
+      return ''
+    })()
+    return chainId === bchChainId ? `bitcoincash:${address}` : address
+  } catch {
+    const accountType = accountMetadata?.accountType
+    const bip44Params = accountMetadata?.bip44Params
+
+    const chainAdapter = getChainAdapterManager().get(chainId)!
+
+    const firstReceiveAddress = await chainAdapter.getAddress({
+      wallet,
+      accountNumber: bip44Params.accountNumber,
+      accountType,
+      index: 0,
+    })
+    return firstReceiveAddress
+  }
 }
