+++ src/pages/Lending/hooks/useLendingQuoteQuery.ts
@@ -10,14 +10,12 @@ import { getReceiveAddress } from 'components/MultiHopTrade/hooks/useReceiveAddr
 import { useDebounce } from 'hooks/useDebounce/useDebounce'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { toBaseUnit } from 'lib/math'
+import { fromThorBaseUnit } from 'lib/utils/thorchain'
 import { selectAssetById } from 'state/slices/assetsSlice/selectors'
 import { selectMarketDataById } from 'state/slices/marketDataSlice/selectors'
 import type { LendingDepositQuoteResponseSuccess } from 'state/slices/opportunitiesSlice/resolvers/thorchainLending/types'
 import { getMaybeThorchainLendingOpenQuote } from 'state/slices/opportunitiesSlice/resolvers/thorchainLending/utils'
-import {
-  BASE_BPS_POINTS,
-  fromThorBaseUnit,
-} from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
+import { BASE_BPS_POINTS } from 'state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils'
 import { selectPortfolioAccountMetadataByAccountId } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
