+++ src/state/slices/opportunitiesSlice/resolvers/thorchainLending/utils.ts
@@ -6,10 +6,11 @@ import { getConfig } from 'config'
 import { type BigNumber, bn, bnOrZero } from 'lib/bignumber/bignumber'
 import type { ThornodePoolResponse } from 'lib/swapper/swappers/ThorchainSwapper/types'
 import { assetIdToPoolAssetId } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
+import { toThorBaseUnit } from 'lib/utils/thorchain'
 import { selectAssetById } from 'state/slices/selectors'
 import { store } from 'state/store'
 
-import { getAccountAddresses, toThorBaseUnit } from '../thorchainsavers/utils'
+import { getAccountAddresses } from '../thorchainsavers/utils'
 import type {
   Borrower,
   BorrowersResponse,
