+++ src/state/slices/opportunitiesSlice/resolvers/thorchainsavers/utils.ts
@@ -2,7 +2,6 @@ import type { AccountId, AssetId, ChainId } from '@shapeshiftoss/caip'
 import {
   avalancheAssetId,
   bchAssetId,
-  bchChainId,
   binanceAssetId,
   bscAssetId,
   btcAssetId,
@@ -14,8 +13,6 @@ import {
   ltcAssetId,
 } from '@shapeshiftoss/caip'
 import type { UtxoBaseAdapter, UtxoChainId } from '@shapeshiftoss/chain-adapters'
-import type { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import { TxStatus } from '@shapeshiftoss/unchained-client'
 import type { Result } from '@sniptt/monads'
 import { Err, Ok } from '@sniptt/monads'
 import axios from 'axios'
@@ -23,17 +20,12 @@ import { getConfig } from 'config'
 import memoize from 'lodash/memoize'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
 import type { Asset } from 'lib/asset-service'
-import type { BN } from 'lib/bignumber/bignumber'
-import { BigNumber, bn, bnOrZero } from 'lib/bignumber/bignumber'
-import type {
-  ThornodePoolResponse,
-  ThornodeStatusResponse,
-} from 'lib/swapper/swappers/ThorchainSwapper/types'
+import { BigNumber, bnOrZero } from 'lib/bignumber/bignumber'
+import type { ThornodePoolResponse } from 'lib/swapper/swappers/ThorchainSwapper/types'
 import { assetIdToPoolAssetId } from 'lib/swapper/swappers/ThorchainSwapper/utils/poolAssetHelpers/poolAssetHelpers'
-import type { AccountMetadata } from 'state/slices/portfolioSlice/portfolioSliceCommon'
-import { isUtxoAccountId, isUtxoChainId } from 'state/slices/portfolioSlice/utils'
+import { fromThorBaseUnit, toThorBaseUnit } from 'lib/utils/thorchain'
+import { isUtxoAccountId } from 'state/slices/portfolioSlice/utils'
 
-import type { getThorchainLendingPosition } from '../thorchainLending/utils'
 import type {
   MidgardPoolPeriod,
   MidgardPoolRequest,
@@ -45,7 +37,7 @@ import type {
   ThorchainSaversWithdrawQuoteResponseSuccess,
 } from './types'
 
-const THOR_PRECISION = '8'
+export const THOR_PRECISION = '8'
 export const BASE_BPS_POINTS = '10000'
 
 export const THORCHAIN_AFFILIATE_NAME = 'ss'
@@ -92,7 +84,7 @@ const SUPPORTED_THORCHAIN_SAVERS_CHAIN_IDS = SUPPORTED_THORCHAIN_SAVERS_ASSET_ID
   assetId => fromAssetId(assetId).chainId,
 )
 
-export const getAccountAddressesWithBalances = async (
+const getAccountAddressesWithBalances = async (
   accountId: AccountId,
 ): Promise<{ address: string; balance: string }[]> => {
   if (isUtxoAccountId(accountId)) {
@@ -149,29 +141,6 @@ export const getAllThorchainSaversPositions = async (
   if (!opportunitiesData) return []
 
   return opportunitiesData
-}
-
-export const getThorchainTransactionStatus = async (txHash: string, skipOutbound?: boolean) => {
-  const thorTxHash = txHash.replace(/^0x/, '')
-  const { data: thorTxData, status } = await axios.get<ThornodeStatusResponse>(
-    `${getConfig().REACT_APP_THORCHAIN_NODE_URL}/lcd/thorchain/tx/status/${thorTxHash}`,
-    // We don't want to throw on 404s, we're parsing these ourselves
-    { validateStatus: () => true },
-  )
-
-  if ('error' in thorTxData || status === 404) return TxStatus.Unknown
-  // Tx has been observed, but swap/outbound Tx hasn't been completed yet
-  if (
-    // Despite the Tx being observed, things may be slow to be picked on the THOR node side of things i.e for swaps to/from BTC
-    thorTxData.stages.inbound_finalised?.completed === false ||
-    thorTxData.stages.swap_status?.pending === true ||
-    (!skipOutbound && thorTxData.stages.outbound_signed?.completed === false)
-  )
-    return TxStatus.Pending
-  if (thorTxData.stages.swap_status?.pending === false) return TxStatus.Confirmed
-
-  // We shouldn't end up here, but just in case
-  return TxStatus.Unknown
 }
 
 export const getThorchainSaversPosition = async ({
@@ -289,24 +258,6 @@ export const getMidgardPools = async (
   return poolsData
 }
 
-export const fromThorBaseUnit = (valueThorBaseUnit: BigNumber.Value | null | undefined): BN =>
-  bnOrZero(valueThorBaseUnit).div(bn(10).pow(THOR_PRECISION)) // to crypto precision from THOR 8 dp base unit
-
-export const toThorBaseUnit = ({
-  valueCryptoBaseUnit,
-  asset,
-}: {
-  valueCryptoBaseUnit: BigNumber.Value | null | undefined
-  asset: Asset
-}): BN => {
-  if (!asset?.precision) return bn(0)
-
-  return bnOrZero(valueCryptoBaseUnit)
-    .div(bn(10).pow(asset?.precision)) // to crypto precision from THOR 8 dp base unit
-    .times(bn(10).pow(THOR_PRECISION))
-    .decimalPlaces(0) // THORChain expects ints, not floats
-}
-
 export const isAboveDepositDustThreshold = ({
   valueCryptoBaseUnit,
   assetId,
@@ -376,50 +327,4 @@ export const makeDaysToBreakEven = ({
   // and ~ a day after deposit
   const daysToBreakEven = BigNumber.max(daysToBreakEvenOrZero, 1).toFixed(0)
   return daysToBreakEven
-}
-
-export const getThorchainFromAddress = async ({
-  accountId,
-  assetId,
-  getPosition,
-  accountMetadata,
-  wallet,
-}: {
-  accountId: AccountId
-  assetId: AssetId
-  getPosition: typeof getThorchainLendingPosition | typeof getThorchainSaversPosition
-  accountMetadata: AccountMetadata
-  wallet: HDWallet
-}): Promise<string> => {
-  const { chainId } = fromAssetId(assetId)
-  if (!isUtxoChainId(chainId)) return Promise.resolve(fromAccountId(accountId).account)
-
-  try {
-    const position = await getPosition({
-      accountId,
-      assetId,
-    })
-    if (!position) throw new Error(`No position found for assetId: ${assetId}`)
-    const address: string = (() => {
-      // THORChain lending position
-      if ('owner' in position) return position.owner
-      // THORChain savers position
-      if ('asset_address' in position) return position.asset_address
-      // For type completeness - if we have a response, we *should* either have an `owner` or `asset_address` property
-      return ''
-    })()
-    return chainId === bchChainId ? `bitcoincash:${address}` : address
-  } catch {
-    const accountType = accountMetadata?.accountType
-    const bip44Params = accountMetadata?.bip44Params
-    const chainAdapter = getChainAdapterManager().get(chainId)!
-
-    const firstReceiveAddress = await chainAdapter.getAddress({
-      wallet,
-      accountNumber: bip44Params.accountNumber,
-      accountType,
-      index: 0,
-    })
-    return firstReceiveAddress
-  }
 }
